+++ config/packages/security.yaml
@@ -18,6 +18,9 @@ security:
             entity: 
                 class: App\Entity\Admin
                 property: username
+        
+    role_hierarchy:
+        ROLE_ADMIN: [ROLE_ADMIN, ROLE_USER]
 
     firewalls:
         dev:
@@ -45,4 +48,5 @@ security:
     # Note: Only the *first* access control that matches will be used
     access_control:
         - { path: ^/admin, roles: ROLE_ADMIN }
-        # - { path: ^/profile, roles: ROLE_USER }
+        - { path: ^/client, roles: ROLE_USER }
+        - { path: ^/phone, roles: ROLE_USER }
