+++ src/Controller/SecurityController.php
@@ -2,20 +2,11 @@
 
 namespace App\Controller;
 
-use App\Entity\User;
 use Firebase\JWT\JWT;
-use App\Service\Token;
-use App\Service\AddUser;
-use App\Service\Content;
-use App\Service\Message;
-use App\Service\Persist;
-use App\Service\AddEntity;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
 class SecurityController extends AbstractController
 {
